+++ app.py
@@ -16,7 +16,7 @@ def main():
 def fill_color():
     color = request.form.get('fill_rgb')
     #print(color)
-    r,g,b = hex2rgb(color)
+    r,g,b = AmbientLEDs.hex2rgb(color)
     print('RGB: {0},{1},{2}'.format(r,g,b))
     ambient_leds.fill(i_red,i_green,i_blue)
     return redirect('/')
