+++ tvleds/__init__.py
@@ -1,2 +1,1 @@
 from .tvleds import AmbientLEDs
-from .utilities import *
